fix(telemetry): record token counts as i64 so spans export integers
tracing-opentelemetry's span attribute visitor implements record_i64
but not record_u64, so u64 token counts fall back to record_debug and
export as string attributes. Phoenix coerces the top-level prompt and
completion counts but drops string-typed prompt_details, so cache
counts were stored yet invisible in the token breakdown. Cast every
gen_ai.usage.* record to i64.

+++ rig/rig-core/src/agent/prompt_request/mod.rs
@@ -466,8 +466,8 @@ where
                 }
 
                 agent_span.record("gen_ai.completion", &merged_texts);
-                agent_span.record("gen_ai.usage.input_tokens", usage.input_tokens);
-                agent_span.record("gen_ai.usage.output_tokens", usage.output_tokens);
+                agent_span.record("gen_ai.usage.input_tokens", usage.input_tokens as i64);
+                agent_span.record("gen_ai.usage.output_tokens", usage.output_tokens as i64);
 
                 // If there are no tool calls, depth is not relevant, we can just return the merged text response.
                 return Ok(PromptResponse::new(merged_texts, usage));
